Rigid_body_tick access and FastDodge
FastDodge now successfully rotates and dodges diagonally.
AerialDodge now accounts for the one-frame gap in jump inputs needed to execute a dodge.
To make this possible, minor restructuring to CowBot and GameState were needed to keep track of whether the bot had jumped on the previous frame.
The result is that CarState now has a jumped_on_last_frame attribute.  Other functionality should be unaffected.