team_sign deprecated in calculations
Instead of calculating how quantities should change based on team, we now always assume we are blue team.
All sign changes take place in GameState.py, and should not be taken into account anywhere else.
x_sign is still used in kickoff code to check if we're on the left or the right side.