Strategy Rework and aerial clears
Reorganized decision making states to be clearer.
Added aerial clear from waiting in net.
So far just when the ball is in the air close to the net - this prevents the "bot dish".
Will add more general saves/clears soon.
Fast Kickoffs in a state of repair, not currently very good.